+++ cisco_sdwan/tasks/implementation/_report.py
@@ -5,7 +5,7 @@ import yaml
 from datetime import date
 from difflib import unified_diff, HtmlDiff
 from typing import Optional, Any, NamedTuple, Annotated
-from collections.abc import Iterator
+from collections.abc import Iterator, Sequence
 from pydantic import field_validator, model_validator, BaseModel, ValidationError, Field, ConfigDict
 from cisco_sdwan.__version__ import __doc__ as title
 from cisco_sdwan.base.rest_api import Rest
@@ -248,10 +248,10 @@ class TaskReport(Task):
     def is_api_required(parsed_args) -> bool:
         return parsed_args.subtask_handler is TaskReport.subtask_create and parsed_args.workdir is None
 
-    def runner(self, parsed_args, api: Optional[Rest] = None) -> list | None:
+    def runner(self, parsed_args, api: Optional[Rest] = None) -> Sequence | None:
         return parsed_args.subtask_handler(self, parsed_args, api)
 
-    def subtask_create(self, parsed_args, api: Optional[Rest]) -> list | None:
+    def subtask_create(self, parsed_args, api: Optional[Rest]) -> Sequence | None:
         source_info = f'Local workdir: "{parsed_args.workdir}"' if api is None else f'SD-WAN Manager URL: "{api.base_url}"'
         self.log_info(f'Report create task: {source_info} -> "{parsed_args.file}"')
 
@@ -286,7 +286,7 @@ class TaskReport(Task):
         return result
 
     # noinspection PyUnusedLocal
-    def subtask_diff(self, parsed_args, api: Optional[Rest]) -> list | None:
+    def subtask_diff(self, parsed_args, api: Optional[Rest]) -> Sequence | None:
         self.log_info(f'Report diff task: "{parsed_args.report_a}" <-> "{parsed_args.report_b}"')
         report_a = Report.load(parsed_args.report_a)
         self.log_info(f'Loaded report "{parsed_args.report_a}"')
@@ -438,7 +438,7 @@ class ReportCreateArgs(TaskArgs):
     @field_validator('file', mode='before')
     @classmethod
     def validate_report_file(cls, v):
-        filename = (isinstance(v, str) and v) or f'report_{date.today():%Y%m%d}.txt'
+        filename = v if isinstance(v, str) and v else f'report_{date.today():%Y%m%d}.txt'
         return validate_filename(filename)
 
     @model_validator(mode='after')